Fix bugs on annotations behaviour, implement new factories.
