+++ modules/annotations/history.js
@@ -286,7 +286,6 @@ window.addEventListener("beforeunload", (e) => {
     _focus(bbox, objectId = undefined, adjustZoom=true) {
         bbox.left = Number.parseFloat(bbox.left || bbox.x);
         bbox.top = Number.parseFloat(bbox.top || bbox.y);
-        if (!Number.isFinite(bbox.left) || !Number.isFinite(bbox.top)) return;
 
         let targetObj = undefined;
         if (objectId !== undefined) {
@@ -294,7 +293,16 @@ window.addEventListener("beforeunload", (e) => {
             if (targetObj) {
                 this.highlight(targetObj);
                 this._context.canvas.setActiveObject(targetObj);
+
+                if (!Number.isFinite(bbox.left) || !Number.isFinite(bbox.top)) {
+                    console.warn("Annotation focus BBOX undefined: try to recompute.");
+                    bbox = targetObj.getBoundingRect(true, true);
+                }
             }
+        }
+
+        if (!Number.isFinite(bbox.left) || !Number.isFinite(bbox.top)) {
+            return;
         }
 
         if (adjustZoom && bbox.width > 0 && bbox.height > 0) {
@@ -401,18 +409,12 @@ else { opener.${_this._globalSelf}._boardItemSave(); } return false;">edit</span
     }
 
     _boardItemEdit(self, focusBBox, object) {
+        let updateUI = false;
         if (this._editSelection) {
             this._boardItemSave(true);
         } else {
-            $('#bord-for-annotations .Box-header').css('background', 'var(--color-merge-box-error-indicator-bg)');
-            this._context.setMouseOSDInteractive(false);
-            this._context.enableInteraction(false);
+            updateUI = true;
         }
-
-        self.parent().find("input").each((e, t) => {
-            $(t).attr('readonly', false);
-        });
-        self.html('save');
 
         let objectId;
         if (typeof object !== "object") {
@@ -422,16 +424,31 @@ else { opener.${_this._globalSelf}._boardItemSave(); } return false;">edit</span
             objectId = object.incrementId;
         }
 
-        this._editSelection = {
-            incrementId: objectId,
-            self: self,
-            target: object
-        };
-
         if (object) {
             let factory = this._context.getAnnotationObjectFactory(object.factoryId);
             if (factory) factory.edit(object);
-        } else console.warn("Object edit: no active object.");
+
+            if (updateUI) {
+                $('#bord-for-annotations .Box-header').css('background', 'var(--color-merge-box-error-indicator-bg)');
+                this._context.setMouseOSDInteractive(false);
+                this._context.enableInteraction(false);
+            }
+
+            self.parent().find("input").each((e, t) => {
+                $(t).attr('readonly', false);
+            });
+            self.html('save');
+
+            this._editSelection = {
+                incrementId: objectId,
+                self: self,
+                target: object
+            };
+
+        } else {
+            this._context.raiseEvent('error', {code: 'NO_OBJECT_ON_EDIT'});
+            console.warn("Object edit: no active object.");
+        }
     }
 
     _boardItemSave(switches=false) {
@@ -442,14 +459,7 @@ else { opener.${_this._globalSelf}._boardItemSave(); } return false;">edit</span
             if (obj) {
                 //if target was set, object could have been edited, update
                 let factory = this._context.getAnnotationObjectFactory(obj.factoryId);
-                let newObject = factory.recalculate(obj);
-                if (newObject) {
-                    this._context.replaceAnnotation(obj, newObject, true);
-                    //from user testing: disable modification of meta
-                    //obj = newObject;
-                } else {
-                    this._context.canvas.renderAll();
-                }
+                factory.recalculate(obj);
             } else {
                 //from user testing: disable modification of meta
                 //obj = this._findObjectOnCanvasById(this._editSelection.incrementId);
