+++ modules/annotations/include.json
@@ -7,8 +7,10 @@
         "annotations.js",
         "convert/convertor.js",
         "convert/asapXml.js",
+        "presets.js",
         "objects.js",
-        "free_form_tool.js",
+        "objectGenericFactories.js",
+        "freeFormTool.js",
         "history.js"
     ],
     "requires": ["webgl", "fabricjs"]
